requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=23.5
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.6
pypdf>=4.0
beautifulsoup4>=4.12
requests>=2.31
lxml>=5.0
playwright>=1.41
rapidfuzz>=3.6
